refactor(main.el): Declutter the code
Improve modularity and cleanup redundancy.

Instead of having multiple `org-ereader-annot--pocketbook-parse-*` and
redundancy in `org-ereader-annot-pocketbook-insert*`, the interaction
between `org-ereader-annot-pocketbook-insert*` and `org-ereader--pocketbook-parse`
is now being handled by `org-ereader-annot--pocketbook-result`.  It receive
boolean argument that defined in `org-ereader-pocketbook-insert*`
to define whether the function intended is to `sort-by-page` or not.

Also checkmark on `README.org`